+++ magda-web-client/src/UI/Pagination.js
@@ -19,7 +19,7 @@ class Pagination extends Component {
         return (
             <button
                 onClick={this.onClick.bind(this, currentIndex - 1)}
-                className="btn-prev au-btn au-btn--secondary"
+                className="btn-prev"
             >
                 {" "}
                 <img src={left_arrow} alt="previous page" />{" "}
@@ -31,7 +31,7 @@ class Pagination extends Component {
         return (
             <button
                 onClick={this.onClick.bind(this, currentIndex + 1)}
-                className="btn-next au-btn au-btn--secondary"
+                className="btn-nexty"
             >
                 {" "}
                 <img src={right_arrow} alt="next page" />{" "}
@@ -40,11 +40,7 @@ class Pagination extends Component {
     }
 
     renderDisabledButton() {
-        return (
-            <button className="au-btn au-btn--secondary" disabled={true}>
-                ...
-            </button>
-        );
+        return <button disabled={true}>...</button>;
     }
 
     renderPageList(max, current) {
@@ -59,8 +55,10 @@ class Pagination extends Component {
                             <li key={i}>
                                 <button
                                     onClick={this.onClick.bind(this, i)}
-                                    className={`au-btn ${
-                                        i === current ? "" : "au-btn--secondary"
+                                    className={`${
+                                        i === current
+                                            ? "current"
+                                            : "non-current"
                                     }`}
                                 >
                                     {i}
@@ -71,8 +69,8 @@ class Pagination extends Component {
                         <li>
                             <button
                                 onClick={this.onClick.bind(this, max)}
-                                className={`au-btn ${
-                                    max === current ? "" : "au-btn--secondary"
+                                className={`${
+                                    max === current ? "current" : "non-current"
                                 }`}
                             >
                                 {max}
@@ -87,10 +85,7 @@ class Pagination extends Component {
                     {current > 1 && this.renderPrevButton(current)}
                     {margins.map(i => (
                         <li key={i}>
-                            <button
-                                className="au-btn au-btn--secondary"
-                                onClick={this.onClick.bind(this, i)}
-                            >
+                            <button onClick={this.onClick.bind(this, i)}>
                                 {i}
                             </button>
                         </li>
@@ -98,18 +93,15 @@ class Pagination extends Component {
                     <li>{this.renderDisabledButton()}</li>
                     <li>
                         <button
-                            className="au-btn"
                             onClick={this.onClick.bind(this, current)}
+                            className="current"
                         >
                             {current}
                         </button>
                     </li>
                     <li>{this.renderDisabledButton()}</li>
                     <li>
-                        <button
-                            className="au-btn au-btn--secondary"
-                            onClick={this.onClick.bind(this, max)}
-                        >
+                        <button onClick={this.onClick.bind(this, max)}>
                             {max}
                         </button>
                     </li>
@@ -124,8 +116,8 @@ class Pagination extends Component {
                         <li key={i}>
                             <button
                                 onClick={this.onClick.bind(this, i)}
-                                className={`au-btn ${
-                                    i === current ? "" : "au-btn--secondary"
+                                className={`${
+                                    i === current ? "current" : "non-current"
                                 }`}
                             >
                                 {i}