+++ CHANGES.md
@@ -6,6 +6,7 @@
 *   Made region and dataset configurable in search api.
 *   Removed underline on coat of arms.
 *   Implemented `Tooltip` component to replace `react-tooltip` from `QualityIndicator`
+*   Made region and dataset configurable in search api.
 
 ## 0.0.39