+++ CHANGES.md
@@ -1,3 +1,5 @@
+## 0.0.41
+
 ## 0.0.40
 
 *   Changed Publishers to Organisations in UI